fix harfbuzz related unit tests

+++ src/mbgl/text/harfbuzz.cpp
@@ -75,9 +75,9 @@ void HBShaper::CreateComplexGlyphIDs(const std::string &text,
     for (uint32_t i = 0; i < glyphCount; ++i) {
         glyphIDs.emplace_back(glyphInfo[i].codepoint, type);
 
-        uint32_t x_advance = glyphPos[i].x_advance / 64;
-        uint32_t x_offset = glyphPos[i].x_offset / 64;
-        uint32_t y_offset = glyphPos[i].y_offset / 64;
+        float x_advance = static_cast<float>(glyphPos[i].x_advance / 64);
+        float x_offset = static_cast<float>(glyphPos[i].x_offset / 64);
+        float y_offset = static_cast<float>(glyphPos[i].y_offset / 64);
 
         adjusts.emplace_back(x_offset, y_offset, x_advance);
     }
@@ -110,9 +110,9 @@ void HBShaper::CreateComplexGlyphIDs(const std::u16string &text,
     for (uint32_t i = 0; i < glyphCount; ++i) {
         glyphIDs.emplace_back(glyphInfo[i].codepoint, type);
 
-        float x_advance = glyphPos[i].x_advance / 64.0;
-        float x_offset = glyphPos[i].x_offset / 64.0;
-        float y_offset = -glyphPos[i].y_offset / 64.0f;
+        float x_advance = static_cast<float>(glyphPos[i].x_advance / 64.0);
+        float x_offset = static_cast<float>(glyphPos[i].x_offset / 64.0);
+        float y_offset = static_cast<float>(-glyphPos[i].y_offset / 64.0f);
 
         adjusts.emplace_back(x_offset, y_offset, x_advance);
     }
